Code clean up. Added ShareApp component to storybook
